+++ src/OpenTelemetry.Exporter.InMemory/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* Changed error handling, `InMemoryExporter` will now throw
+  `ObjectDisposedException` if `Export` is invoked after the exporter is
+  disposed.
+  ([#3607](https://github.com/open-telemetry/opentelemetry-dotnet/issues/3607))
+
 ## 1.4.0-alpha.2
 
 Released 2022-Aug-18